+++ Common/Services/Integrations/ExpandedStorage/ExpandedStorageIntegration.cs
@@ -1,5 +1,6 @@
 namespace StardewMods.Common.Services.Integrations.ExpandedStorage;
 
+/// <inheritdoc />
 internal sealed class ExpandedStorageIntegration : ModIntegration<IExpandedStorageApi>
 {
     private const string ModUniqueId = "furyx639.ExpandedStorage";