+++ CustomBush/ModEntry.cs
@@ -6,6 +6,7 @@ using StardewModdingAPI.Events;
 using StardewMods.Common.Interfaces;
 using StardewMods.Common.Services;
 using StardewMods.Common.Services.Integrations.FauxCore;
+using StardewMods.CustomBush.Framework;
 using StardewMods.CustomBush.Framework.Models;
 using StardewMods.CustomBush.Framework.Services;
 
@@ -16,6 +17,10 @@ public sealed class ModEntry : Mod
 
     /// <inheritdoc />
     public override void Entry(IModHelper helper) => this.Helper.Events.GameLoop.GameLaunched += this.OnGameLaunched;
+
+    /// <inheritdoc />
+    public override object GetApi(IModInfo mod) =>
+        new CustomBushApi(this.container.GetInstance<BushManager>(), mod, this.container.GetInstance<ILog>());
 
     private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
     {
@@ -35,7 +40,7 @@ public sealed class ModEntry : Mod
         this.container.RegisterInstance(this.Helper.ModRegistry);
         this.container.RegisterInstance(this.Helper.Reflection);
         this.container.RegisterInstance(this.Helper.Translation);
-        this.container.RegisterInstance<Func<Dictionary<string, BushModel>>>(this.GetData);
+        this.container.RegisterInstance<Func<Dictionary<string, CustomBush>>>(this.GetData);
         this.container.RegisterSingleton<AssetHandler>();
         this.container.RegisterSingleton<BushManager>();
         this.container.RegisterSingleton<IEventManager, EventManager>();
@@ -48,10 +53,10 @@ public sealed class ModEntry : Mod
         this.container.Verify();
     }
 
-    private Dictionary<string, BushModel> GetData()
+    private Dictionary<string, CustomBush> GetData()
     {
         var assetHandler = this.container.GetInstance<AssetHandler>();
         var gameContentHelper = this.container.GetInstance<IGameContentHelper>();
-        return gameContentHelper.Load<Dictionary<string, BushModel>>(assetHandler.DataPath);
+        return gameContentHelper.Load<Dictionary<string, CustomBush>>(assetHandler.DataPath);
     }
 }